requirements_demo.txt
scikit-image
packaging
monai
opencv-python==4.6.0.66
pandas
scikit-learn
matplotlib
h5py
yacs
timm
huggingface_hub
# For DeepGaze baseline
numpy
pillow

# For Keras-based wrappers
tensorflow==2.12.0            # or tensorflow-macos/tensorflow-metal if on macOS with MPS
h5py
matplotlib                     # for plt.imsave used in server response
